// FBOLinx.Web/FBOLinx.Web/ClientApp/src/app/models/fees-and-taxes.ts
export type FeesAndTaxes = {
    calculationType: number;
    departureType: number;
    fboid: number;
    flightTypeClassification: number;
    isOmitted: boolean;
    name: string;
    oid: number;
    omitsByCustomer: any;
    omitsByPricingTemplate: any[];
    value: number;
    whenToApply: number;
};